pyproject.toml
[tool.poetry]
name = "programming-bot"
version = "0.1.0"
description = ""
authors = ["0xWaleed <imwaleed@outlook.sa>"]
readme = "README.md"
packages = [{include = "programming_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
poe-api = "^0.3.1"

[tool.poetry.scripts]
bot="programming_bot.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

